Move tile-specific constant to tile list constants module
